作業前にterminalのコードを\repos\discord\command\terminal\common.pyに移動
