+++ py-builder/repos/discord/command/terminal/common.py
@@ -1,0 +1,18 @@
+#!ignore
+from ....imports import *
+from ....constant import *
+from ....logger.logger_create import *
+from ....minecraft.read_properties import *
+from ....assets.text_dat import *
+from ....assets.utils import *
+#!end-ignore
+
+#!open ./repos/discord/command/terminal/set.py
+#!ignore
+from .set import *
+#!end-ignore
+#!open ./repos/discord/command/terminal/delete.py
+#!ignore
+from .delete import *
+#!end-ignore
+
